+++ quickcheck-classes-base/src/Test/QuickCheck/Classes/Eq.hs
@@ -4,22 +4,26 @@
 
 module Test.QuickCheck.Classes.Eq
   ( eqLaws
+  , substitutiveEqLaws
   ) where
 
 import Data.Proxy (Proxy)
 import Test.QuickCheck hiding ((.&.))
 import Test.QuickCheck.Property (Property)
+import Test.QuickCheck.Function
 
 import Test.QuickCheck.Classes.Internal (Laws(..))
 
 -- | Tests the following properties:
 --
 -- [/Transitive/]
---   @a == b ∧ b == c ⇒ a == c@
+--   @a '==' b ∧ b '==' c ⇒ a '==' c@
 -- [/Symmetric/]
---   @a == b ⇒ b == a@
+--   @a '==' b ⇒ b '==' a@
 -- [/Reflexive/]
---   @a == a@
+--   @a '==' a@
+-- [/Negation/]
+--   @x '/=' y '==' 'not' (x '==' y)@
 --
 -- Some of these properties involve implication. In the case that
 -- the left hand side of the implication arrow does not hold, we
@@ -48,3 +52,21 @@ eqSymmetric _ = property $ \(a :: a) b -> case a == b of
 
 eqReflexive :: forall a. (Show a, Eq a, Arbitrary a) => Proxy a -> Property
 eqReflexive _ = property $ \(a :: a) -> a == a
+
+eqNegation :: forall a. (Show a, Eq a, Arbitrary a) => Proxy a -> Property
+eqNegation _ = property $ \(x :: a) y -> (x /= y) == not (x == y)
+
+-- | Tests the following properties:
+--
+-- [/Substitutive/]
+--   @x '==' y ⇒ f x '==' f y@
+--
+-- /Note/: This does not test `eqLaws`.
+-- If you want to use this, You should use it in addition to `eqLaws`.
+substitutiveEqLaws :: forall a. (Eq a, Arbitrary a, CoArbitrary a, Function a, Show a) => Proxy a -> Laws
+substitutiveEqLaws _ = Laws "Eq"
+  [ ("Substitutivity"
+    , property $ \(x :: a) y (f :: Fun a Integer) ->
+        x == y ==> applyFun f x == applyFun f y
+    )
+  ]